+++ anime_ml/settings.py
@@ -1,7 +1,7 @@
 import logging
 import os
 
-logging.basicConfig(level=logging.NOTSET, format=f"%(asctime)s - %(levelname)s - %(message)s")
+logging.basicConfig(level=logging.INFO, format=f"%(asctime)s - %(levelname)s - %(message)s")
 
 BASE_URL: str = "https://api.myanimelist.net/v2"
 AUTH_URL: str = "https://myanimelist.net/v1/oauth2/authorize"